Added jdbc to mysql data type mapping feature
Mostly, the feature auto-mapped the jdbc data type to mysql data type
depending on the data_type value (from  java.sql.Types) and the mysql
data type's max length, which is customized in
JdbcDataTypeToMySqlMapping.xml.
As we added the feature, we have to updated some data types' constructor
as some types have to be constructed without display length. Besides,
constructors with primitive type arguments have to use the encapsulated
types as arguments because of reflection's need.